adjustments on himeno

+++ tasks/mopp-2018-t3-himeno/himeno.h
@@ -2,7 +2,8 @@
 #define __HEADER_HIMENO__
 
 #include "matrix.h"
+#include "common.h"
 
-float jacobi( unsigned int n, Matrix* M1, Matrix* M2, Matrix* M3, Matrix* M4, Matrix* M5, Matrix* M6, Matrix* M7 );
+double jacobi( uint n, Matrix* M1, Matrix* M2, Matrix* M3, Matrix* M4, Matrix* M5, Matrix* M6, Matrix* M7 );
 
 #endif
